pyproject.toml
[project]
name = "mandelbrot-mpi"
version = "1.0.0"
description = "Large Scale Modeling Project 1"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "numpy>=2.0.0",
    "matplotlib>=3.9.0",
    "mpi4py>=3.1.0",
    "mlflow>=2.15.0",
    "pyyaml>=6.0",
    "pytest>=8.0",
    "pandas>=2.2.0",
    "ipykernel>=7.0.1",
    "mpi-pytest>=2025.7",
    "pytest-mpi>=0.6",
    "snakeviz>=2.2.2",
    "ruff>=0.14.1",
    "pyarrow>=21.0.0",
    "numba>=0.62.1",
    "seaborn>=0.13.2",
    "scienceplots>=2.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mandelbrot"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
